+++ appbuilder/__init__.py
@@ -108,8 +108,8 @@ from appbuilder.utils.logger_util import logger
 
 from appbuilder.core.utils import get_model_list
 
-from appbuilder.core.console.app_builder_client.agent_builder import AppBuilderClient
-from appbuilder.core.console.app_builder_client.agent_builder import AgentBuilderClient
+from appbuilder.core.console.appbuilder_client.appbuilder_client import AppBuilderClient 
+from appbuilder.core.console.appbuilder_client.appbuilder_client import AgentBuilder
 
 from .core._exception import (
     BadRequestException,
@@ -177,5 +177,5 @@ __all__ = [
     "ImageUnderstand",
     "MixCardOCR",
     "AppBuilderClient",
-    "AgentBuilderClient",
+    "AgentBuilder",
 ]
